refactor(footer): rename component, fix responsive breakpoint and standardize icons 'react icons'
