+++ src/components/Footer/index.module.scss
@@ -55,6 +55,7 @@
       list-style-type: none;
       padding: 0;
       margin: 0;
+      margin-top: 1.1rem;
     }
 
     a {
@@ -132,9 +133,31 @@
 }
 
 // -=- RESPONSIVIDADE -=-
-// DESKTOP/TABLET
+// DESKTOP
 
-@media (max-width: 1400px) {
+@media (max-width: 1440px) {
+  .footer {
+    padding: 3.5rem 5rem;
+  }
+
+  .flexFooter {
+    // flex-direction: column;
+    gap: 3rem;
+  }
+
+  .containerFooterLinks {
+    justify-content: space-between;
+
+    nav {
+      padding: 0 5rem;
+    }
+  }
+}
+
+// -=- RESPONSIVIDADE -=-
+// TABLET
+
+@media (max-width: 1024px) {
   .footer {
     padding: 4rem 4rem;
   }
